Refactor add_resource_files() to allow resource files to be added asynchronously
